Main Game Loop
Testing out code design idea in a seperate scene. Idea is to abstract out a lot of the functionality which will hopefully speed up development. Seems to be working right now so going to keep working on it to show some examples as to why we should do it this way