Pause working with some extra
Loop idea seems to be working ok. Was easily able to implement a pause menu which works, stops characters from recieveing inputs and animations will now pause and resume once the game unpauses. Also built out a template for enemies, which I expect to change to fit with how Matt wishes to create enemies, really just an example of how it will work with the loop. Starting to build a basic UI, with textmesh pro, to make testing easier, that we can expand and build on later, and developing the players other systems like their 2 weapons and bars etc.