New Player model and animations & game over screen
Replaced the old player and camera prefab with the new one player model. Combos are broken right now due to the changes and may cause the players animator to become "stuck", just sorting that out now and adding the new attack animations from tests done last night, but everything else as far as I am aware works.

Also the player can now die. If they do a small menu appears and gives the player the option to restart. by pressing a button, which reloads the scene. Just temp for the time being.