Checkpoints
Added the first iteration of the checkpoints system. If the player collides with the checkpoint it will become the new checkpoint that the player spawns at with when they respawn after a death. Check points cannot be recollided with, as there isn't any backtracking,  so they will always be sent to the furthest checkpoint they got, but it is easy enough to change.